update feature favourite in SearchFragment

+++ app/src/main/java/com/example/iti_project/ui/RecipeActivity/Favourit/FavoriteFragment.kt
@@ -2,6 +2,7 @@ package com.example.iti_project.ui.RecipeActivity.Favourit
 
 import android.app.ProgressDialog
 import android.os.Bundle
+import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
@@ -53,30 +54,17 @@ class FavoriteFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        viewModel.getFavoriteList()
+//        viewModel.getFavoriteList()
         return inflater.inflate(R.layout.fragment_favourit, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         rv_favoriteRecipes = view.findViewById(R.id.rv_favorite_recipe)
-
-    }
-
-    override fun onResume() {
-        super.onResume()
-        GlobalScope.launch(Dispatchers.Main) {
-            delay(20000)
-            instantiateProductsRecyclerView()
-
-            favoriteRecipesAdapter.setData(
-                viewModel.favoriteRecipes.value as MutableList<Meals>,
-                viewModel.favoriteUserIds
-            )
-            listenToDeleteFavouriteItems()
-        }
-//        listenToUpdateFavouriteItems()
-
+//        viewModel.getFavoriteList()
+        instantiateProductsRecyclerView()
+        listenToDeleteFavouriteItems()
+        listenToUpdateFavouriteItems()
     }
 
     private fun instantiateProductsRecyclerView() {
@@ -94,24 +82,29 @@ class FavoriteFragment : Fragment() {
 
 
     private fun listenToDeleteFavouriteItems() {
-        favoriteRecipesAdapter.favoriteUserRemovedIds.observe(viewLifecycleOwner) { response ->
-            if (!response.isNullOrEmpty()) {
-                viewModel.deleteFavoriteRecipe(response)
-                viewModel.getFavoriteList()
+        if (favoriteRecipesAdapter.favoriteUserRemovedIds != null) {
+            favoriteRecipesAdapter.favoriteUserRemovedIds.observe(viewLifecycleOwner) { response ->
+                if (!response.isNullOrEmpty()) {
+                    viewModel.deleteFavoriteRecipe(response)
+//                    viewModel.getFavoriteList()
 //                favoriteRecipesAdapter.setData(viewModel.favoriteRecipes , viewModel.favoriteUserIds)
+                }
             }
         }
     }
 
     private fun listenToUpdateFavouriteItems() {
+
+        Log.i("response", "  " + "nooooooooooo")
         viewModel.favoriteRecipes.observe(viewLifecycleOwner) { response ->
+            Log.i("response", "  " + response.first().strMealThumb)
             if (!response.isNullOrEmpty()) {
+
                 favoriteRecipesAdapter.setData(
-                    response as MutableList<Meals>,
-                    viewModel.favoriteUserIds
+                    response as MutableList<Meals>
                 )
-//                favoriteRecipesAdapter.setData(viewModel.favoriteRecipes , viewModel.favoriteUserIds)
             }
         }
+
     }
 }
